mod-bq-findings/application/requirements.txt
google-cloud-dlp==3.9.2
google-cloud-storage==2.5.0
google-cloud-pubsub==2.13.10
google-cloud-datastore==2.9.0
google-cloud-bigquery==3.3.5